[core] Move `nodeFactory` to its own module
